requirements.txt
pymessenger==0.0.7.0
gunicorn==19.6.0
watson-developer-cloud==1.0.1
django==1.11
django-redis-sessions==0.6.1
djangorestframework==3.6.3
django-heroku==0.2.0
